+++ app/src/main/java/ru/eruditeonline/app/data/repository/ResultRepository.kt
@@ -18,7 +18,7 @@ interface ResultRepository {
 
     /** Сохранение результата */
     suspend fun saveResult(
-        completeId: Long,
+        completeId: Int,
         name: String,
         surname: String,
         patronymic: String?,
